Moved SVG Icon to Docs
