+++ src/docs/DocsNavigation/DocsSidebar.svelte
@@ -2,7 +2,7 @@
 	import { page } from '$app/stores';
 	import { writable, type Writable } from 'svelte/store';
 
-	import SvgIcon from '$lib/components/SvgIcon/SvgIcon.svelte';
+	import SvgIcon from '$docs/SvgIcon/SvgIcon.svelte';
 	import { menuNavLinks } from './links';
 	import AppRail from '$lib/components/AppRail/AppRail.svelte';
 	import AppRailTile from '$lib/components/AppRail/AppRailTile.svelte';
