+++ src/docs/SvgIcon/SvgIcon.svelte
