+++ src/docs/SvgIcon/SvgIcon.test.ts
